% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdc.R
\name{fdc}
\alias{fdc}
\title{Exceedance Probability Equation}
\usage{
fdc(infile)
}
\arguments{
\item{infile}{a numeric list or vector input file}
}
\value{
A dataframe with variable and probability
}
\description{
This function estimates the flow duration curve. It assumes the input is
an atomic vector containting values of a variable i.e. discharge
}
